Break the build in case patches cannot be applied
Currently the build doesn't fail in case patches are broken or no longer
valid. This change tries to mitigate this issue. However the way patches
are currently handled and applied is error prone and needs more rework,
e.g. deleting packages before every build is just a workaround because
there is no proper way to clean them up. There must be a better way to
deal with patches, which do not change gluon code directly, but OpenWRT or
package code.