Increase delay before batctl throughput_override
On some devices it may take more than a second, before a new interface is registered in BATMAN, so increasing the timeout helps to avoid wired mesh, where throughput is not set correctly. However using "sleep" is a bad idea, there should be a better hook to set the throughput override, when BATMAN is ready to, but I couldn't find it yet.