Fix overriding the main git remote by accident
In case the gluon-build folder exists and does not contain a .git directory (or
file) the following command "git remote set-url origin" will override the git
remote of the "main" folder/directory.

Avoiding this by checking explicitely for the .git directory inside gluon-build
instead of just checking for the folder's existence.