Put target list into a separate file
As the Github build is building targets in separate VMs, it does not use the target-list in the Makefile, but holds its own copy instead. This change moves the target list into a new file called "targets" which is then used by the Makefile as well as the Github workflow.