domains: merge domains for parker, switch config server
This merges all domains to one for each community, plus ffwert_events:
- ffmuc
- ffdon
- ffwer
- ffwert_events

They all have new domain seeds and new mesh ids, to avoid interference with non-parker nodes.

Before a rollout we need to add a function to migrate the domain settings on all nodes.

Also change the config_server to pgw01.ext.ffmuc.net (until we have a proper setup with our webfrontend proxies).

And last but not least, change the DNS server to the parker gateway IP running DNS64 - needs to be extended when new gateways are setup.